requirements.txt
Django==1.7
argparse==1.2.1
django-model-utils==2.2
django-phonenumber-field==0.6
phonenumbers==6.3.0
wsgiref==0.1.2
